moved monkeypatched weather source to conftest.py
