+++ tests/modeling/test_hourly_load_profile_model.py
@@ -16,35 +16,6 @@ from eemeter.modeling.exceptions import (
 )
 from eemeter.structures import EnergyTrace
 from eemeter.modeling.models import HourlyLoadProfileModel
-from eemeter.weather import WeatherSource
-
-
-def _fake_temps(usaf_id, start, end, normalized, use_cz2010):
-    # sinusoidal fake temperatures in degC
-    dates = pd.date_range(start, end, freq='H', tz=pytz.UTC)
-    num_years = end.year - start.year + 1
-    n = dates.shape[0]
-    avg_temp = 15
-    temp_range = 15
-    period_offset = - (2 * np.pi / 3)
-    temp_offsets = np.sin(
-        (2 * np.pi * num_years * np.arange(n) / n) + period_offset)
-    temps = avg_temp + (temp_range * temp_offsets)
-    return pd.Series(temps, index=dates, dtype=float)
-
-
-@pytest.fixture
-def monkeypatch_temperature_data(monkeypatch):
-    monkeypatch.setattr(
-        'eemeter.weather.eeweather_wrapper._get_temperature_data_eeweather',
-        _fake_temps
-    )
-
-
-@pytest.fixture
-def mock_isd_weather_source():
-    ws = WeatherSource('722880', False, False)
-    return ws
 
 
 @pytest.fixture
